UPnP Port Forwarding Release
Ever wanted to create a server for a game so you and your friends can play together, but couldn't figure out port-forwarding and didn't want to install certain programs that are a little sketchy and/or aimed at a different audience... I've got you!

This port-forwarding tool was created in Python utilizing the miniupnpc as the core and tkinter for a usable interface - I will provide both the code and any other files needed in my little pocket of the internet, additionally, I will have a build provided as a .exe (which Windows will be unhappy if you download and run it, but I am not a verified developer) which is done using pyinstaller so you don't need to separately install Python and this code's dependencies.

I appreciate feedback and will maintain this from time to time.